rework step
