+++ .github/workflows/black.yml
@@ -11,4 +11,4 @@ jobs:
         with:
           options: "--check --diff"
           src: "./sciopy"
-          version: "23.1.0"
+          version: "24.10.0"
